remove the non-asi-version of sort, rename the asi-version of sort to sort, add mozart_env for configuring the environment
