--- accelerators/stratus_hls/asi_sort_stratus/hw/hls/Makefile
@@ -1,1 +1,0 @@
-../../../common/hls/Makefile
