--- accelerators/stratus_hls/asi_sort_stratus/hw/sim/.gitignore
@@ -1,1 +1,0 @@
-sort
