--- accelerators/stratus_hls/asi_sort_stratus/hw/sim/Makefile
@@ -1,3 +1,0 @@
-# Copyright (c) 2011-2022 Columbia University, System Level Design Group
-# SPDX-License-Identifier: Apache-2.0
-include ../../../common/systemc.mk
