--- accelerators/stratus_hls/asi_sort_stratus/hw/sort.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<sld>
-  <accelerator name="sort" desc="SORT float[32:1024][1:1024]" data_size="4" device_id="00b" hls_tool="stratus_hls">
-    <param name="len" desc="input vector length (number of float)"/>
-    <param name="batch" desc="number of input vectors"/>
-    <param name="len_min" value="32"/>
-    <param name="len_max" value="1024"/>
-    <param name="batch_max" value="1024"/>
-    <param name="prod_valid_offset" desc="prod_valid_offset" />
-    <param name="prod_ready_offset" desc="prod_ready_offset" />
-    <param name="cons_valid_offset" desc="cons_valid_offset" />
-    <param name="cons_ready_offset" desc="cons_ready_offset" />
-    <param name="input_offset" desc="input_offset" />
-    <param name="output_offset" desc="output_offset" />
-  </accelerator>
-</sld>
