--- accelerators/stratus_hls/asi_sort_stratus/hw/src/sort_debug_info.hpp
@@ -1,9 +1,0 @@
-// Copyright (c) 2011-2022 Columbia University, System Level Design Group
-// SPDX-License-Identifier: Apache-2.0
-
-#ifndef __SORT_DEBUG_INFO_HPP__
-#define __SORT_DEBUG_INFO_HPP__
-
-typedef uint32_t debug_info_t;
-
-#endif // __SORT_DEBUG_INFO_HPP__
