--- accelerators/stratus_hls/asi_sort_stratus/sw/baremetal/Makefile
@@ -1,4 +1,0 @@
-# Copyright (c) 2011-2022 Columbia University, System Level Design Group
-# SPDX-License-Identifier: Apache-2.0
-APPNAME := sort
-include $(DRIVERS)/common_bare.mk
