--- accelerators/stratus_hls/asi_sort_stratus/sw/linux/app/Makefile
@@ -1,5 +1,0 @@
-# Copyright (c) 2011-2022 Columbia University, System Level Design Group
-# SPDX-License-Identifier: Apache-2.0
-EXTRA_CFLAGS ?=
-APPNAME := sort
-include $(DRIVERS)/common.mk
