--- accelerators/stratus_hls/asi_sort_stratus/sw/linux/driver/Kbuild
@@ -1,3 +1,0 @@
-EXTRA_CFLAGS += \
-	     -I$(DRIVERS)/include -I$(DRIVERS)/../common/include -I$(ACC_SW)/linux/include
-obj-m := sort_stratus.o
