--- accelerators/stratus_hls/asi_sort_stratus/sw/linux/driver/Makefile
@@ -1,5 +1,0 @@
-include $(DRIVERS)/driver.mk
-
-Module.symvers:
-	$(MAKE) -C $(ESP_CORE_PATH) CROSS_COMPILE=$(CROSS_COMPILE) ARCH=$(ARCH)
-	cp $(ESP_CORE_PATH)/$@ .
