+++ accelerators/stratus_hls/sort_stratus/hw/hls/project.tcl
@@ -1,4 +1,4 @@
-#  Copyright (c) 2011-2023 Columbia University, System Level Design Group
+#  Copyright (c) 2011-2022 Columbia University, System Level Design Group
 #  SPDX-License-Identifier: Apache-2.0
 
 ############################################################
@@ -73,20 +73,40 @@ define_system_module tb ../tb/system.cpp ../tb/sc_main.cpp
 ######################################################################
 set DEFAULT_ARGV "1024 16"
 
-foreach dma [list 32 64] {
-    define_io_config * IOCFG_DMA$dma -DDMA_WIDTH=$dma
+# Baseline acc - no ASI
+foreach dma [list 64] {
+    define_io_config * IOCFG_DMA$dma\_BASELINE -DDMA_WIDTH=$dma
 
-    define_system_config tb TESTBENCH_DMA$dma -io_config IOCFG_DMA$dma
+    define_system_config tb TESTBENCH_DMA$dma\_BASELINE -io_config IOCFG_DMA$dma\_BASELINE
 
-    define_sim_config "BEHAV_DMA$dma" "sort BEH" "tb TESTBENCH_DMA$dma" -io_config IOCFG_DMA$dma -argv $DEFAULT_ARGV
+    define_sim_config "BEHAV_DMA$dma\_BASELINE" "sort BEH" "tb TESTBENCH_DMA$dma\_BASELINE" -io_config IOCFG_DMA$dma\_BASELINE -argv $DEFAULT_ARGV
 
     foreach cfg [list BASIC] {
-	set cname $cfg\_DMA$dma
-	define_hls_config sort $cname -io_config IOCFG_DMA$dma --clock_period=$CLOCK_PERIOD $COMMON_HLS_FLAGS -DHLS_DIRECTIVES_$cfg
+	set cname $cfg\_DMA$dma\_BASELINE
+	define_hls_config sort $cname -io_config IOCFG_DMA$dma\_BASELINE --clock_period=$CLOCK_PERIOD $COMMON_HLS_FLAGS -DHLS_DIRECTIVES_$cfg
 	if {$TECH_IS_XILINX == 1} {
-	    define_sim_config "$cname\_V" "sort RTL_V $cname" "tb TESTBENCH_DMA$dma" -io_config IOCFG_DMA$dma -argv $DEFAULT_ARGV -verilog_top_modules glbl
+	    define_sim_config "$cname\_V" "sort RTL_V $cname" "tb TESTBENCH_DMA$dma\_BASELINE" -io_config IOCFG_DMA$dma\_BASELINE -argv $DEFAULT_ARGV -verilog_top_modules glbl
 	} else {
-	    define_sim_config "$cname\_V" "sort RTL_V $cname" "tb TESTBENCH_DMA$dma" -io_config IOCFG_DMA$dma -argv $DEFAULT_ARGV
+	    define_sim_config "$cname\_V" "sort RTL_V $cname" "tb TESTBENCH_DMA$dma\_BASELINE" -io_config IOCFG_DMA$dma\_BASELINE -argv $DEFAULT_ARGV
+	}
+    }
+}
+
+# With ASI
+foreach dma [list 64] {
+    define_io_config * IOCFG_DMA$dma\_SM -DDMA_WIDTH=$dma -DENABLE_SM
+
+    define_system_config tb TESTBENCH_DMA$dma\_SM -io_config IOCFG_DMA$dma\_SM
+
+    define_sim_config "BEHAV_DMA$dma\_SM" "sort BEH" "tb TESTBENCH_DMA$dma\_SM" -io_config IOCFG_DMA$dma\_SM -argv $DEFAULT_ARGV
+
+    foreach cfg [list BASIC] {
+	set cname $cfg\_DMA$dma\_SM
+	define_hls_config sort $cname -io_config IOCFG_DMA$dma\_SM --clock_period=$CLOCK_PERIOD $COMMON_HLS_FLAGS -DHLS_DIRECTIVES_$cfg
+	if {$TECH_IS_XILINX == 1} {
+	    define_sim_config "$cname\_V" "sort RTL_V $cname" "tb TESTBENCH_DMA$dma\_SM" -io_config IOCFG_DMA$dma\_SM -argv $DEFAULT_ARGV -verilog_top_modules glbl
+	} else {
+	    define_sim_config "$cname\_V" "sort RTL_V $cname" "tb TESTBENCH_DMA$dma\_SM" -io_config IOCFG_DMA$dma\_SM -argv $DEFAULT_ARGV
 	}
     }
 }
