+++ accelerators/stratus_hls/sort_stratus/hw/sim/Makefile
@@ -1,3 +1,3 @@
-# Copyright (c) 2011-2023 Columbia University, System Level Design Group
+# Copyright (c) 2011-2022 Columbia University, System Level Design Group
 # SPDX-License-Identifier: Apache-2.0
 include ../../../common/systemc.mk
