+++ accelerators/stratus_hls/sort_stratus/hw/sort.xml
@@ -6,5 +6,11 @@
     <param name="len_min" value="32"/>
     <param name="len_max" value="1024"/>
     <param name="batch_max" value="1024"/>
+    <param name="prod_valid_offset" desc="prod_valid_offset" />
+    <param name="prod_ready_offset" desc="prod_ready_offset" />
+    <param name="cons_valid_offset" desc="cons_valid_offset" />
+    <param name="cons_ready_offset" desc="cons_ready_offset" />
+    <param name="input_offset" desc="input_offset" />
+    <param name="output_offset" desc="output_offset" />
   </accelerator>
 </sld>
