+++ accelerators/stratus_hls/sort_stratus/hw/src/sort_debug_info.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2011-2023 Columbia University, System Level Design Group
+// Copyright (c) 2011-2022 Columbia University, System Level Design Group
 // SPDX-License-Identifier: Apache-2.0
 
 #ifndef __SORT_DEBUG_INFO_HPP__
