+++ accelerators/stratus_hls/sort_stratus/sw/baremetal/coh_func.h
