+++ accelerators/stratus_hls/sort_stratus/sw/linux/app/Makefile
@@ -1,4 +1,4 @@
-# Copyright (c) 2011-2023 Columbia University, System Level Design Group
+# Copyright (c) 2011-2022 Columbia University, System Level Design Group
 # SPDX-License-Identifier: Apache-2.0
 EXTRA_CFLAGS ?=
 APPNAME := sort
