+++ accelerators/stratus_hls/sort_stratus/sw/linux/app/coh_func.h
