+++ mozart_env
@@ -1,0 +1,21 @@
+# HLS settings
+export VIVADO=/PATH/TO/YOUR/VIVADO/  # E.g. /PATH/TO/YOUR/Vivado/2019.2/
+export STRATUS_PATH=/PATH/TO/YOUR/STRATUS/ # E.g. /PATH/TO/YOUR/cadence-Mar2020/STRATUS192
+
+export XILINX_VIVADO=$VIVADO
+source $VIVADO/settings64.sh
+export CDS_AUTO_64BIT=all
+export PATH=$PATH:$VIVADO/bin/
+
+# RISC-V
+export RISCV=/PATH/TO/YOUR/RISCV  # E.g. /PATH/TO/YOUR/riscv
+export PATH=$PATH:/PATH/TO/YOUR/RISCV/BIN  # E.g. /PATH/TO/YOUR/riscv/bin
+
+# Leon3
+export PATH=$PATH:/PATH/TO/YOUR/LEON3/BIN # E.g. /PATH/TO/YOUR/leon/bin
+export PATH=$PATH:/PATH/TO/YOUR/LEON3/MKLINUXIMG  # E.g. /PATH/TO/YOUR/leon/mklinuximg
+export PATH=$PATH:/PATH/TO/YOUR/LEON3/SPARC-ELF/BIN # E.g. /PATH/TO/YOUR/leon/sparc-elf/bin
+
+export XILINXD_LICENSE_FILE=YOUR.XILINX.LICENCE
+export ESP_ROOT=$PWD
+
